Update to central eterna

+++ src/main/java/me/hapyl/fight/game/heroes/doctor/PhysGun.java
@@ -91,7 +91,7 @@ public class PhysGun extends Weapon {
         }
 
         @Override
-        public Cooldown setCooldownSec(int cooldownSec) {
+        public Cooldown setCooldownSec(float cooldownSec) {
             return super.setCooldownSec(cooldownSec);
         }
     }
